Try to find a real temp directory for temporary files
Until now maxima_tempdir was normally set to maxima_userdir.

This was bad for the following reasons:
 - on windows gnuplot isn't able to handle files from directories that
   contain non-ASCII characters. Since maxima_userdir normally reads
   c:\users\<username> if the username is not plain ASCII this would
   hinder plots from working.
 - If the user's home directory was on a network drive plotting also
   had problems due to this drive being somehow less accessible.
 - The home directory was littered with temporary files and
 - on android maxima_userdir therefore had to be set to a temporary
   folder hindering the user from placing a maxima-init.mac file.

I have added many fallbacks to the new logic. If they are still not
sufficient for any use-case I beg to add the necessary code to this
file.

Tested on Win7 and Ubuntu; Resolves the following wxMaxima issues:
 - https://github.com/andrejv/wxmaxima/issues/664
 - https://github.com/andrejv/wxmaxima/issues/732