[bugs:#3208] algsys.lisp:presultant: resimplify result
This is a partial fix to [bugs:#3208]. It is necessary but not
sufficient to fix the problem.

$resultant is called on an expression.  The result is equivalent to 0.
The return value was not simplified and therefore was non-zero.
Consequently, all valid solutions are rejected.

A simple example is:
  ?trace(presulatant);
  ?trace(?$resultant);
  eq3:676*y^2-4056*y-28*3^(7/2)+2007$
  eq4:676*y^2+20*3^(5/2)-333$
  resultant(eq3,eq4,y);
  algsys(eq3,eq4],[y]);

No changes to testsuite or contrib_ode testsuite (a heavy user of algsys).
Minimal change in testsuite runtimes.