Make run_testsuite(share_tests=true) work
When share_tests=true, the file search path was not set correctly
because file_search_tests was not included.  Hence, the regular test
files couldn't be found.

If the test file can't be found, we print out a message to that effect
so the user has some idea about what happened instead of an
uninformative error break message.

Also took this opportunity to add a bit more debugging info and small
cleanup.