Fix #3221: elliptic_pi(n,z,m) wrong for z > %pi/2
Use the properties of elliptic_pi to reduce the range of z to the
interval [0, %pi].  Then the interval [%pi/2, %pi] can be converted to
the interval [0, %pi/2].  Thus, we can use Carlson's RJ function to
compute all of the necessary terms.

A simple derivation of the algorithm is also included.  A bigfloat
version is also implemented.