Revert "Renamed solver.mac to Solver.mac"
Revert because this commit clobbered the file which actually
contained the working code (Solver.mac) with the one which
contained nothing but a message + load the other one.

So let's restore the working code. I'll nuke the trivial solver.mac
(so that we don't have two files which are the same name in
case-insensitive systems) in another commit.

This reverts commit 6620590e9cfc702af4a71d1c511e2aabe4ce6730.