Remove an unused lexical variable in DEF%TR for MDO
The lexical variable TEM is unused.  The code which used TEM
previously was removed in commit d72823 and was only for Lisp
Machines.  I'm thinking that this leftover lone TEM in the body
was originally there to silence a warning about TEM being unused
when not on a Lisp Machine (but this is just a guess).