Revert Commit [7c85cf]
Seems unnecessary and does break VTK on Windows.
Sorry for that.
Still unsure, if the path to the generated file should be
quoted somehow - what if the path (which contains the username
and may contain special chars, spaces, unicode characters.

At least partly - kept the #+(or windows win32 win64)
(not sure if it is necessary).