Unset the variable LANGUAGE in maxima.bat.in
It is usually not set on Windows, but inherited from the Linux/Unix
environment, if Wine is using for testing the Windows build.

And it *does* cause troubles, at least with Clisp, maybe other Lisps too.
The commit solves errors like:

*** - MAKE-PATHNAME: Illegal :DIRECTORY argument
      (:RELATIVE "de_AT:de" "LC_MESSAGES")
The following restarts are available:
MACSYMA-QUIT   :R1      Maxima top-level

when starting Maxima/Clisp in Wine.