plotdf UI is not consistent with use of ':' in dy/dx
This is another patch from themusicgod1. It was handed in via
https://bugs.launchpad.net/ubuntu/+source/maxima/+bug/1670191/
The original description was:

tl;dr typo in Xmaxima menu to configure plotdf / plotdf "Plot Setup"

process to replicate:
1) load("plotdf");
2) plotdf(x,[x,-1,1],[y,-1,1]);
3) hit the wrench icon to get into the config menu
4) switch between dy/dx and dy/dt,dx/dt

note that the : in dy/dx and dy/dt changes from ":" to ""
if you start in dy/dt,dx/dt dy/dt is shown as "dy/dt:", then if you switch to dy/dx and back it's shown as "dy/dt".
if you start in dy/dx dy/dx is shown as "dy/dx"

this would all be consistent if the variables dy/dx, and dy/dt are consistent with other variables in this config menu and if they are listed with ':' which this patch should do.