Documentation: A introduction to simplification.
We definitively need a short introduction to this chapter as its contents
currently isn't sufficient for explaining an user how this very important
aspect of maxima works. Richard Fateman and Roland Salz have made big
efforts in writing such a thing with suggestions from the maxima mailing
list => adding it to the TeXinfo manual. If there are ideas on how to
further improve this section please edit it or send in patches. But I
think it provides a good base for understanding this central point of
what to expect from wxMaxima.