operatingsystem: Make chdir work with recent gcl versions.
The fix was found by Thorsten Bonow. His comment was:

I think the bug is caused because in newer versions of gcl, chdir
is still defined in the si package, but isn't exported---meaning that
it does not belong to the public part of the API of the package which
is accessible from other packages by calling it with the package name
and a one colon qualifier.  Using two colons makes it possible to
access any symbol in a package.