Prevent local property leaks when a block has an invalid variable list
When a block had an invalid variable list, an error would occur and
MUNLOCAL would not be called (presumably because local could not have
been called in the block's body at this point).  The problem is that
LOCLIST was modified before these error checks were done, so these
errors would leave a superfluous NIL in LOCLIST.

If the block was evaluated at top-level, this extra NIL would be left
over in LOCLIST.  If the block was evaluated inside a function then
this all caused the cleanup of local properties to be one frame behind.
An error in nested blocks would make it multiple frames behind if the
error was in an inner block.

(%i1) foo () := (local (f), f () := bar, block ([1]))$

(%i2) foo ();
Only symbols can be bound; found: 1
#0: foo()
 -- an error. To debug this try: debugmode(true);

(%i3) f ();
(%o3) bar

Now move the LOCLIST modification to after these error checks.

The test suite runs fine.  As with commit 9387d6, I'm not sure how
to add a test for this since we need to cause an error, but using
errcatch prevents the leak.