Tried to make all warnings use a consistent format.
The idea of this step is to make automatic detection of warnings
by a gui frontend easier and less prone to produce false positives
on messages like "0 warnings, 0 errors".

As warnings began with " warning:", "Warning: ", "Warning - ",
"<MODULENAME>: Warning:", Warning- <Modulename>- " there was the
possibility to pick one or two of the existing formats.
The following two were selected:

 - Warning: <message text>
 - <module name>: Warning: <message text>

The next nightly build of wxMaxima now will look for the following
two regExps in order to find warnings and automatically format them
accordingly:

 - "^Warning: "
 - ": Warning: "