Documentation: Tried to make the entry for "for" more novice-friendly.
Novices normally need loads of time before they determine that if the
result of a loop should be a list "makelist" is the right command.

Most other manual entries have the form

   Head-Description-See also-Examples.

Here the description is accompanied by examples. Also this manual entry
is long enough that not everybody will reach the traditional location
 => Did add the cross-ref at the head of the entry.

Other changes:
 - Omitted the line that tells that the elements of the for loop can
   be shuffled around a bit as I agreed with the commentary that this
   line isn't necessary
 - Moved the general form of loops to a block at the beginning of the
   entry (which should make things earlier for the hasty that feel to
   just need one example)
 - Added a link to the chapter "performance considerations for lists"
   (it is easy to create O(n^2) complete constructs using for, after
   all)
 - Reformulated the first paragraphs a bit.

Hope it was OK not to split this into 8 separate commits (one for each
change) as I originally planned: This approach felt like flooding the
repo with commits this time and - after all only one manual entry was
changed.