Share testbench: Resolve a recurring clisp failure on launchpad
clisp sometimes fails the share testbench on launchpad. The reason
is that clisp isn't too fast and commit 532fa16bb0 that besides
others provides "additional regression tests for rtest_hg" added a
test that sees if hg is fast enough.

=> Increased the timeout by a factor of 10 so even people with an
1st generation intel atom processor should be able to pass the
test.

An additional question now could be if the test still makes
sense after its limit has been increased so much and if there are
even slower computers around that still fail the test => if that
test really makes sense. But there is more than one answer to
this kind of question, normally.