wrstcse: use a rectangular, not a gauss-shaped random distribution
First feedback from the users: If we use a gauss-shaped random
distribution when choosing random values between the minimum and
the maximum value of a parameter this means that the parameters
we try rarely exceed the interval [minimum,maximum].

Also using linear random distributions allows to demonstrate that
If a system depends on a big number of parameters the result of
the system will automatically show a reasonably good gaussian
distribution.

Using a linear random distribution also allows to drop the
dependency on the distrib package which is quite big.