Default settings for gcl that allow the testbench to succeed
In its default setup gcl tries to save time by running the garbage
collector as seldom as possible - which can cause memory fragmentation
or running out of memory before the garbage collector is actually run.

It also defaults to run the garbage collector only if all gcl instances
together near 50% of the physical memory installed in the current
computer. The new default settings gc more aggressively. But if firefox
is started while maxima runs the 50% of memory usage still might result
in an OOM.

I didn't change the defaults for MSW as we currently don't provide
maxima using gcl on windows and gcl might change before we need
the changed defaults.