More multitasking-friendly default settings for gcl
By default all gcl instances try to share 50% of the physical memory
installed in the current machine which is partially fine as this
reduces the time spent garbage-collecting and allows big machines to
use more memory for this. But if gnome-shell, firefox and thunderbird
are running at the same time it at least causes loads of swapping.