Check for installed maxima flavour before falling back to MAXIMA_DEFAULT_LISP
The nightly linux build is split into maxima-gcl, maxima-sbcl, maxima-ecl,
maxima-clisp and in the future maybe others. If maxima-sbcl isn't
installed currently but was the default at compile time we don't want this
to produce an error but want this to cause maxima to use one of the
installed flavours instead.