Fix some shell quoting issues (maxima-local, xmaxima-local, src/maxima)
Hopefully this takes care of all issues with spaces in the pathname,
but we'll see!

xmaxima-local.in:
Add double-quotes in various places so it works with spaces in the
pathname.

maxima-local.in:
Add double-quotes around the argument to dirname *and* the argument
to cd so it works with spaces in the pathname.  Commit f8e1509f fixed
most places where spaces in the pathname could cause problems, but
this was skipped (my guess is because something like ./maxima-local
still worked).  Besides fixing this just to fix it, this is also
needed so things like ./xmaxima-local work when there are spaces in
the absolute pathname.

src/maxima.in:
Add double-quotes around the argument to dirname *and* the argument
to cd so it works with spaces in the pathname.

[The exact way these scripts would fail when spaces were in the
pathname varied from system to system, depending on how dirname and
cd behave when given more than one argument.  Since I've already spent
some time looking into this, here is some information for those who
are curious:

Many Bourne-compatible (or "mostly" Bourne-compatible) shells have a
built-in cd that gives an error when given more than one argument,
but at least ksh and zsh have a special substitution behavior when
given two arguments.

The dirname that appears on some BSDs gives an error when given more
than one argument (and in the scripts above this causes cd to change
to your home directory, which in turn sets top_srcdir to your home
directory!).  This dirname behavior seems to conform to POSIX.1.

GNU dirname and the dirname that appears on some other BSDs behave by
basically outputting the dirname for each argument on separate lines
(but in the scripts above this couldn't work as desired in general
because then cd would fail to properly change the directory since it
would be given more than one argument).  On some of these systems the
man pages say that this dirname behavior is "compatible" with POSIX.2
(I don't know, but I'm guessing that that simply means this behavior
is not forbidden).]