Add a script to test loading packages twice or package combinations.
As suggested on the mailing list on 2016-12-01
(see https://sourceforge.net/p/maxima/mailman/message/35526552/)

This script loads packages
- once
- twice
- any combination of tow packages.
and generate a report for these commands.

(A 'package' is currently a .mac file in the share directory.
Yes, Lisp files should also be considered. And some mac files
are probably helper files for real packages and should not be loaded
by itself. But without a proper naming scheme (what is an
'helper' file, what is an real package) that can not be done.

Demo files, test files (because the computation time they need)
and files which require interactive inputs are ignored.

The test run for package combinations runs very long.