Fix bug #2116: lambda form for taylor_simplifier
The problem was that the value of taylor_simplifier had to be a
symbol because it was passed to FBOUNDP in TAYLOR1.  This was noted
in a comment in src/hayat.lisp but not in the documentation.

I think the idea is that (for efficiency) if the value of
taylor_simplifier is an fbound symbol then it should just be
funcalled directly (in various places in src/hayat.lisp) rather than
funcalling TAYLOR_SIMPLIFIER_CALLER, which calls MCALL, which calls
MAPPLY1, and so on.

Now an explicit check is done for a symbol before calling FBOUNDP on
the value of taylor_simplifier.

The test suite runs fine, with new tests.