Remove an old, commented-out ERRSET macro definition
This ERRSET was present in Maxima v5.0 and was used on old lisps
which had no error handling (or at least when there was no other
ERRSET definition available).  All it did was evaluate the first
argument and wrap its return value in a list.

The ERRSET macro using HANDLER-CASE has been used for all lisps
(and this old definition has been commented-out) since commit
ad12e542 in 2004.