Fix trace_break_arg assignment in nested breakpoints
This is to fix a bug introduced in commit 721457f3.  The idea was
to handle a particular corner case involving a traced function that
returns multiple values, the break trace option, an exit breakpoint
and the assignment of trace_break_arg to the same value it started
with.  But I didn't correctly handle cases like nested breakpoints
("nested" in the sense that a new breakpoint is started from within
this exit breakpoint), so some things didn't work as they did before
in this case.