Add a new rtest_trace file to the test suite
This is for testing the functionality of Maxima's trace and timer.

I'm moving over the trace and timer tests I added to rtest_rules
in commit 721457f3 to this new file.  It was trace and timer that
were causing problems there -- not the rules themselves.

The test suite runs fine.