Bug fix: simplode on a singleton list could return a non-string
If simplode was called on a singleton list and an empty delimiter,
then it just returned the list element without converting it to a
string first.

Now the result of calling sconcat on the list element is returned
in this case.

The test suite and share test suite run fine, with a new test.