errcatch: save and print the error message in the case of lisp errors
Now if an otherwise uncaught lisp error occurs during an errcatch,
the error variable will be a list that has error report message
(as a string) as its only element, and the error message will be
printed if errormsg is true.  This is consistent with how Maxima
errors are saved and printed (in an errcatch).

Having this information saved and printed is useful in and of
itself, but it's particularly important for the test suite runs
in the current nightly build setup.  errcatch is used in the test
suite and when a lisp error would occur we would only see
"error-catch" with no information about the error.

The test suite and share test suite run fine.