Remove a superfluous quoted symbol from an AND form
The result of the AND form is only used for a boolean test.  Because
of this and the position of the quoted symbol, I think that maybe
something like (TYPEP POWER 'FIXNUM) was being changed to
(FIXNUMP POWER), but this was leftover.  Either way, this has been
present since at least as far back as Maxima v5.0.

No functional changes.