Fix bug #3427: gcfactor(3+4*%i) => error
The error that occurred was due to a failed assertion.  This commit
removes that assertion.

That assertion was only added in commit 6ef2b26c, which rewrote
GCFACTOR.  The code for the assertion was present before that,
but the assertion did not take place unless the non-standard feature
CL was present on the features list of the underlying lisp (I tested
SBCL, ECL, ABCL and CLISP, but none of these have CL on their
features list).  That commit removed the read-time conditional so
that the assertion always took place.

The test suite and share test suite both run fine, with a new test.