Fix bug #3365: rat(1/3),modulus:6 gives incomprehensible error
This improves the error message in CRECIP and CRECIP-GENERAL when
we cannot take the inverse of the given integer with respect to a
given modulus.

I'm using the wording suggested by Stavros Macrakis in the bug
report.  It's generally much clearer than the current message and
I can't think of any obvious improvements to it.

The new error message is (arguably) still not completely clear in
some cases.  An example is calling rat(1/3) with modulus:3 because
CRECIP is given 0, and so the error message mentions 0 (not 3).