Bug fix: SET-MODULUS allowed setting modulus to 0
Having modulus set to 0 causes massive breakage.

SET-MODULUS is supposed to be a safe way to set modulus (with checks
to validate the new value), but there has been an off-by-one error
since it was introduced in commit 99d0c62d as a rewrite of
SETQMODULUS.

(It wasn't mentioned in the commit message for the aforementioned
commit, but it appears that SETQMODULUS was broken because it
allowed any positive number.  SET-MODULUS fixed that part to only
allow integers, but it was broken because of the incorrect lower
bound.)

After a cursory glance it doesn't appear that a user could get
modulus set to 0 (without using lisp directly), but I didn't bother
doing any in-depth analysis.