Documentation: "memoizing functions" causes less confusion than "array functions"
We originally described 5 types of list-like elements:
 - lists
 - declared arrays
 - undeclared arrays [synonymous to hash arrays]
 - hash arrays and
 - array functions

This caused confusion with the users and many questions.

=> now that the use of memoizing functions is documented
more throughly all that was left was to change their name
from "array functions" to "memoizing functions" and to
hyperlink their name to their description.

This means we now have:
 - lists
 - declared arrays
 - hash lists and
 - memoizing functions

Additionaly corrected a few errors that resulted in confusing
names of things that had near name cashes before the change.

If lists one day are merged with declared arrays as the mailing
list discusses currently we are now already prepared for that.