Somewhat better error messages
Try to print out the args of the function when printing out error
messages. We can only do this if there are no optional args for the
function.  If there are optional args, just print the function name.

Also include the actual arguments of the function when printing out
the error message.