Make the testsuite tell which lisp the current test run is for
We have several build servers that generate Nightly Builds and run
the test bench, two of them being dautermann.at and launchpad.net.

The first step for dealing with them was marking all tests that
failed there as "known bad" so we can see if a commit breaks
something for a lisp that is still used in the wild as the build
servers either use  current lisp versions or ones that are used
in currently active versions of Ubuntu. As some tests occur only
sporadically and as testing is a side-effect of packaging and
packaging a non-c++ project needs non-standard measures making
that work required months.

The next (and potentially more long-running whilst being less
time-consuming) step will be finding out which tests occur only
in old lisp versions. With several caveats:

 - Our existing build servers already show that some test failures
   occur only on some computers (Intel/AMD? 32/64 Bit? Signal
   integrity problems? Weird virtual machines? Does all depend
   on the amount of free RAM, perhaps by triggering the gc
   in different places?)
 - Some lisp versions never change their version number even if
   they change the code.
 - And some Ubuntu versions use updated lisps without telling
   the lisp compiler to use a new version number
=> the 2nd step might not be the final one.