ECL only: defsystem seems not to support autoconf-variables from the build tree
If in src/maxima.system I replace

                     (:file "autoconf-variables")))

by

                     (:private-file "autoconf-variables")))

all lisps I have tested load the file autoconf-variables from the
current directory except from ECL that somehow chokes on the variables
not being defined.

=> made "autoconf-variables" a ":file" again for ECL until we find out
what is the cause: This breaks out-of-source-builds for ECL.
But at least it makes normal builds work for ECL again.