Do not include the build directory for crosscompiling in git.
This directory must now be created before. Updated the README.txt
to reflect these changes (and mentioned ABCL).

Reason: The maxima source code is copied during the build procedure
(I do not use the recently introduced out-of-source build (now) and for
ABCL it does not work anyways) and crosscompile-windows/build is excluded
during the copy process.
And this results always in a 'dirty' git version. (which can be seen for
example as a directory like "share/maxima/branch_5_42_base_159_g105c58f_dirty".
If it is not under version control, this problem should be fixed.