Use old defmfun approach for internal functions
If the function doesn't begin with a $, we use the old defmfun
approach.  Based on some preliminary tests, this makes the test suite
run about 1-2% slower than before.  I think that's a good trade-off.

This is a temporary hack.  We should just replace all of those
internal functions to use defun instead of defmfun.