Ensure "No unexpected errors" is printed
Make sure we print that when there are no unexpected errors, even if
there are tests that unexpected passed.  This is needed by make check
to verify things worked.  So, now always print that and then print out
the list of tests, if any, that passed but were not expected to pass.