"make check": A more canonical way to do the interactive tests
"make check" automatically does the interactive tests as well as
the normal testsuite ones. Until now in order to make this happen
the cons() command was necessary. But in the meantime the list of
test files has changed to be an ordinary file => use append() in
order to add the manual tests.