Switched the default lisp for the windows installer to "clisp" again
Both TCL and wxWidgets have gone faster which means it is easy now to
trigger the clisp bug that causes clisp to send garbled data (data
packets arriving in the wrong order or skipped or extra chars) over
the network to the front-end.

The problem is more visible in wxMaxima as maxima and wxMaxima
communicate using XML and errors are likely to produce XML errors
under these circumstances. But the result might be worse in Xmaxima
as here communication is in ASCII which means wrong but plausible
equations might be output without any warning for the user.