Hide the fact that we autogenerate src/share-subdirs.lisp from the autotools
The autotools delete all autogenerated files they know about on
"make distclean" which means that cleaning up after a build doesn't
return the project to the state it was in before the build. This

 - causes "make distcleancheck" to fail
 - risks that someone deletes the autogenerated file in the repo
   which has happened quite a few times till now either accidentally
   or because the file clearly is autogenerated.

Now we first create the file under a different name and then copy
it over the one we ship with maxima without telling the autotools
about it.