"make check": Tried to make more lisp errors produce useful output
Currently on launchpad's nightly build "make check" on ecl often
results in no useful output as the test is killed after 6 hours
and doesn't output useful data until it is killed. Let's try to
add a few (quit) on stdin in order to recover from a lisp error
and to make the lisp quit - which makes the testsuite output text.