Rename LISPM-MFUNCTION-CALL-AUX to MFUNCTION-CALL-AUX
The LISPM- part of the name was introduced when writing this function
while there was still a copy of the original Maclisp MFUNCTION-CALL-AUX
around that was incompatible.  Later this Maclisp one was removed and
only the LISPM- one remained.

The current name adds no value for us, so let's rename it back to the
shorter original name.  An upcoming commit will also be changing this
function's signature anyway.

No problems with the test suite, share test suite or rtest_translator.