Add back old MFUNCTION-CALL MACRO warning message
This message used to be printed in Macsyma when MFUNCTION-CALLing a
lisp macro.  The printing of this particular message was lost when
converting this code away from Maclisp, and I'm guessing that it was
not intentional.  I think this case deserves a warning message just
as much as the other cases.

No problems with the test suite, share test suite or rtest_translator.