Require gcl 2.6.14 or later
Updates the check so that we require gcl 2.6.14 or later instead of
2.6.13.

This also cleans up some code.  First, remove the version check that
was done when the path to gcl was given explicitly.  The version will
be checked later anyway.

Second, take advantage of the gcl 2.6.14 having variables giving the
major, minor, and extra version numbers.  Then we don't have to do the
heuristic of extracting the version number from the
`lisp-implementation-version`.