Wrap the list of command line options if needed.
When the list of equivalent command line options is long, wrap it so
it looks neat.  This is similar to how we wrap the description of the
option and is based on the same format string.

Of course, when producing the texi output, we don't want to wrap the
options.