In the linespoints style option for plot2d, emit nothing when pointinterval is not specified, instead of emitting " pointinterval 10".
The Gnuplot documentation does not specify what is the default for
pointinterval, although some experimentation suggests that the default
is equivalent to "pointinterval 1". Be that as it may, I have chosen
to emit no specification for pointinterval at all on the general
principle that if the user did not say anything about it, then plot2d
should let Gnuplot do whatever it does when the user does not mention
some option or the other.