move code in display2d-unicode-enabled to adjust-character-encoding
Commit ff883f32a moved and modified CLISP-specific code from
ADJUST-CHARACTER-ENCODING in src/init-cl.lisp to
DISPLAY2D-UNICODE-ENABLED in src/displa.lisp. The intention is to
ensure that the terminal encoding is set to utf-8 correctly.

That created a bug in how CLISP+MAXIMA treated STDOUT when piped
through `tee' [1].

This change moves the setting of CLISP's custom:*terminal-encoding*
from DISPLAY2D-UNICODE-ENABLED to ADJUST-CHARACTER-ENCODING and bind
custom:*default-file-encoding* to the terminal encoding.

This change fixes the tee-related bug.

Refs:
[1] https://sourceforge.net/p/maxima/mailman/message/58815918