src/mload.lisp: default answers_from_file to true
Since the introduction of tests/rtest_ask1.mac, the testsuite needs to
read answers from the interactive tests from the file. That is, the
`run_testsuite' option `answers_from_file' should default to `true'.

This patch makes that change and documents the change.

The documentation of `batch_answers_from_file' is also updated, to
explain where this flag is set to `true' by default.

The Texinfo file doc/info/commandline-options.texi needed to be
referenced, so I have added an anchor, a category for it. A short
comment in doc/info/Command.tex has been added to provide a
dis-ambiguating link to the section on command-line options.

Thanks to Wolfgang Dautermann.

Ref:
https://sourceforge.net/p/maxima/mailman/maxima-discuss/thread/CAAsY_sSB%3DBxUaNbax9EM6Ltdbh5V7MuFtvxHwKO2JXBvXwTZTw%40mail.gmail.com/#msg58831614