src/mload.lisp: move special declaration of $batch_answers_from_file
Previously, the special declaration for $batch_answers_from_file was
placed at the beginning of the function definition.

It needs to be inside the let* binding of $batch_answers_from_file in
order for that binding to be dynamic instead of lexical.

Source: http://clhs.lisp.se/Body/d_specia.htm

With this change, commit 08ba2dcfb works as expected: the
value of $batch_answers_from_file is bound to the value of
ANSWERS_FROM_FILE and TEST-BATCH sees that value.

Ref: https://sourceforge.net/p/maxima/mailman/maxima-discuss/thread/CAG14z1G_pSJuqHcT9EghidgWy2pstSmFx%3DSL5xMdqnD7rY0WXQ%40mail.gmail.com/#msg58834887