Fix #4404:  Clarify error message about incorrect syntax
To make the error message a little easier to read, the keywords are
quoted like using single quote and backquote.

So now we have:

(%i1) for i in [1, 2, 3, 4] thru 3 do print (i);
incorrect syntax: A `do' cannot have a `thru' with a `in' field.
n [1, 2, 3, 4] thru
                  ^

(%i1) for i = 0 upto 7 do print(i);
incorrect syntax: `=' is an unknown keyword in a `do' statement.
for i =
     ^
(%i1) for i thru 3 thru 7 do print (i);
incorrect syntax: The `thru' slot for `do' is already filled.
for i thru 3 thru
                ^

Note that we changed the phrasing for the last error message.  I think
this is clearer than the original when quoting is added.

All the changes were done in the function collision-check.  There are
other places where we should use the quoting.  I didn't check or
change those places.