Configure sets up variables for highlighting
Because the texinfo API changed after 6.8, we have configure figure
out which API to use since it knows which version of makeinfo is being
used.

Specifically, texinfo 6.8 used '$options->{"EXTRA_HEAD"}' to configure
the extra items to be placed in the HTML HEAD element.  But after 6.8,
it changed to the "official" API that uses
'texinfo_set_from_init_file("EXTRA_HEAD", '

This requires updating texi2html.init.in to have these values
substituted in.

Also, the cssSelector needed by highlightJs differs between 6.8 and
later.  For 6.8, we need "pre.example".  But later it's
".example-preformatted".

I tested this with 6.8 and 7.1.  Highlighting works.