Need to define TEXI_EXTRA_TAIL to close off texi setting command
When we're using the new texinfo API that uses the function
`texinfo_set_from_init_file`, we need a closing paren to close off the
function call.  When the we're not using the new API, we don't need to
do anything.

This requires updating configure.ac and texi2html.init.in to deal with
these changes.

This means you need to run bootstrap and configure again to get these
changes.