maxima.bat.in: replace hard encoded lisp executable name with @xxx_NAME@.
That was already commited by Robert Dodier in commit 1a340520a.
I reverted it (partly) later because it did not work when compiling
a crosscompiled installer (commit f9f3a9760).

But it is probably better to use these variables by default
and patch maxima.bat later when building the crosscompiled
Windows installer.
So if someone compiles Maxima (native) on Windows (does anyone
do this?) and uses a special path to a Lisp, it should work as expected.

Corrected the code to change "\" to "/" in the paths; it seems the
slash must be written twice. Don't know why, but it works now. And
a double slash in a path wouldn't hurt anyways.