Windows installer: Update TCL/TK to 9.0.1.
wish86s.exe changed to wish90s.exe.

Had to use the compile option --enable-zipfs=no, as it
does not work with zipfs enabled, but as this seems to be
a new feature, there should be no problems.
Submitted a issue for TCL, maybe I can include that feature
in future:
https://core.tcl-lang.org/tcl/tktview/cb338c130b8fba479c2848de4fcd2045c1db12b1