rtest4: in tests for SF bug #2905, invert case of all-upper and all-lower symbols, namely D --> d, r --> R, z --> Z, in order to make this a valid test for commit fc2272b.
Commit bd2722, which changed the ordering of all-upper and all-lower
symbols, made the bug in this example disappear, but changing the symbol
names, so that they sort into the same order as the original example,
makes the bug reappear (when tested with a post-bd2722 and pre-fc2272b
version).

Note that bd2722 did not change the ordering for mixed-case symbols,
so "Some" is unchanged in this example.