Add some comments on how psi[s](p/q) is computed.
For specific values of psi[s](p/q) that maxima knows how to compute,
give text formulas and references for the formulas.  Include some
derivations if the references don't quite match how they're being
computed.

This of course begs the question on why they're done this way.  It
certainly makes it harder to follow.  But that's a question for
another day.