Fix bug #4458 (?great can mark unsimped expr as simped)
Calling orderlessp(A, B) with unsimplified A could modify A in place, marking
it as simplified, when it's not. This was due to the function SIMPLN1, which is
called by ORDMEXPT, unconditionally passing T as the second argument to
SIMPLIFYA. This told SIMPLIFYA that the arguments are already simplified.
Under certain circumstances, this could result in the behavior described in the
bug report #4458.

SIMPLN1 now checks for the existence of the 'SIMP flag and tells SIMPLIFYA.