Fix COMPARE-BIGFLOATS, use existing functions
Use BIGFLOATP to adjust both bigfloats to the same precision (the bigger one),
then use FPDIFFERENCE for subtracting them. Finally, return the signum of the
mantissa of the difference.

A local macro is used to extract the bigfloats' precisions. There is also the
function BIGFLOAT-PREC, but I could not get it inlined in SIMP.LISP (order of
file compilation?).