Optimize ALPHALESSP (no consing)
This change optimizes the ALPHALESSP function, which is used a lot during
simplification. When comparing two symbols, e.g. $A and $B, the previous
implementation would convert both symbols to a string with modified Lisp
readtable and print case and then compare the strings. The new implementation
doesn't loops over the symbol names' characters and doesn't allocate any memory.

It reduces the time required for (SIMPLIFYA '((MPLUS) $A $B $C) T) by 60% on SBCL.