Optimize $member and $assoc
$member and $assoc would perform operations on the whole list before starting
the actual search, causing operations on long lists to be slow even if the sought
element was at the front. (see bug #4550)

These operations are now done inside the loop, while iterating over the list.

For $assoc, the error message is now a bit clearer, indicating the (first)
offending element instead of printing the whole list, which could be very long.