Optimize ALIKE1 by taking advantage of CAR sharing
Many expressions in Maxima have CARs that are EQ. This is done by the EQTEST
function. For bigfloats, this is done by BCONS.

ALIKE1 previously didn't take advantage of that, always comparing the CARs by
looking at their content. A simple EQ test before doing that speeds up ALIKE1
significantly in cases where the expressions have a shared CAR (majority of cases).

For general expressions, the speed-up is up to 11%.
For bigfloats, the speed-up is up to 33%.
(SBCL 2.5.4)