Add link to Carlson's doc and some cleanups and examples
Add a link to Carlson's "Numerical Computation of Real or Complex
Elliptic Integrals" to each description of Carlson's functions.

We also clarify some text about the E function.  It's not super clear
so we changed it to say that it's the complete elliptic integral of
the second kind and add a link to `elliptic_ec`.

Finally, we add an equation that relates the integral definition of
the elliptic integral of the third kind to Carlson's functions.  Also
add a link to `elliptic_pi` and a note that this integral differs from
`elliptic_pi`.