More debugging
Just do an `ls` of the doc/info dir to see what's there.  Likewise for
tests.

And for some reason when we build with gcl, it looks like we're
actually building with ecl for some reason.  Don't know why, but added
a `--with-gcl` to set the program name, just in case.