Add RMS method support to wrstcse.mac
wrstcse.mac by default tries out all combinations of input tolerances
in order to determine what the absolute maximum of the tolerance of the
output of an equation might be.

The new functions now assume that the range of -1...1 for tol[n] represents
a gauss distribution of the width wc_defaultsigma and try to find tolerances
with a given confidence interval using the RMS method.

Advantages of the new method:
 * Avoids overengineering by allowing to drop cases that are so extreme that they
   will appear only once in a billion samples.
 * Fast, even if many tolerance-ridden parameters have to be regarded.

Disadvantages:
 * It is not extreme worst case
 * It works only where the RMS method is applicable
 * User documentation for the new functions will follow in a later commit.