Make the background for pygments to match examples
We want to background color for pygments highlighting to match the
background used for examples.  Otherwise it looks weird.

To do this, define a CSS variable `--example-background-color` that is
set to the example background color.  Use this variable to set the
example background and the div highlight class for pygments.