Fix bugs in testing value of syntax highlighting
Fix issues with checking for pygments when syntax highlighting was set
to pygments.  We were producing configure errors if
--enable-syntax-highlighting was not set.  If
--enable-syntax-highlighting was set to something else (or not given),
we produced an error that pygmentize was not found.  That was a
mistake in the AS_IF.

Finally, print out the summary for the syntax highlighting correctly
instead of silly things like "syntax highlighting using false".