Fix #4601: regression in spherical_bessel_j integral
When removing the special var `*updn`, we made a mistake in
`ptimes%e`.  The second cond clause in `ptimes%e` was updating the
variable `term` with the result of the (recursive) call to `ptimes%e`.

However, in the original code, we didn't actually update `term`.  We
only used the result to determine if the second clause applied; the
result was discarded.

In addition, for the original code, `*updn` **was** updated, so we
need to make sure we still update `updn` from this call.

Finally, the last clause did a throw and returned NIL.  To be safe (I
think), we should return multiple values, returning `updn` as the
second value.

Added the integral as a test in rtest15.mac.