Allow highlightjs to use github-dark when in dark mode.
For dark mode with highlightjs, we use github-dark (since we're using
github for light mode).

This is done by adding a new link tag to
HIGHLIGHT1 in configure.ac to specify github-dark when dark mode is
preferred.  We also update the existing link to github to specify
light mode for this.

This doesn't look very good in dark mode, though because the default
example background is a dark gray.

But this is a step in supporting dark mode when using highlightjs for
the examples.