Rework highlightjs to make it easier read/modify
The variable HIGHLIGHT1 has a lot of redundant stuff.  Let's define
some variables to make reading and modifying HIGHGLIGHT1 simpler.

* CDN si the main part of the URL where we're loading
  highlightjs from.
* HIGHLIGHT_VER is the version we want to use
* LIGHTMODE is the CSS file we want for light mode display
* DARKMODE is the CSS file we want for dark mode display

Finally, use these variables to define the value of HIGHLIGHT1.  To do
this we use `cat` with a `HERE` document to make shell substitution
easier.  Otherwise we get into a huge mess of shell-quoting issues
with single- and double-quotes.