Update give-up function in def-simplifier
To support a def-simplifier for cabs, we need to update the `give-up`
function to specify the name of noun to use when we give up.  The
old simplifier would do an `eqtest` with `mabs` instead of the
expected `%cabs`.

To make this more general, change `give-up` to take keyword args
instead of the previous optional arg.