If custom-pygments, load up separate CSS files for the styles
If we're using custom-pygments, load up the CSS files for the light
and dark CSS styles to be used with pygments highlighting.

This is done by adding a new variable CUSTOM_PYGMENTS_CSS to
configure.ac to add appropriate HTML link attributes.  Then update
texi2html.init.in to recognize CUSTOM_PYGMENTS_CSS so that the links
are included.