Update version check for makeinfo
Looks like a very recent texinfo HEAD reports it's version as 7.2.90,
so we need to update the test from version 7.2.99 to 7.2.90 as the
version that supports custom highlighting commands.

Also, we weren't printing out the version correctly; we left of the
dot separating the minor version from the patch version.

Finally, comment an echo that is not relevant from the code that
detects the version.