Use a table for the values of expintrep
This makes it easier to see the various possible values for expintrep
with the corresponding explanation.  Also makes linke to these easier
to find.

Note that the anchor names do *not* match the name of the value
because some of the values are also names of functions.  The
references to these values use the new anchor names.